Refactor authentication and task management structure
- Consolidated service registrations in Program.cs by adding AddTaskFlowServices method.
- Updated repository methods in TaskRepository and UserRepository for improved readability.
- Refactored import paths in frontend components to align with new directory structure.
- Removed obsolete auth guard, interceptor, and service files to streamline authentication handling.
