+++ backend/TaskFlow.API/Program.cs
@@ -1,56 +1,10 @@
-using TaskFlow.Model.Repositories;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-using System.Text;
+using TaskFlow.API.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.AddControllers();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-builder.Services.AddScoped<TaskRepository>();
-builder.Services.AddScoped<UserRepository>();
-
-var jwt = builder.Configuration.GetSection("Jwt");
-var signingKey = jwt.GetValue<string>("SigningKey")
-    ?? throw new InvalidOperationException("Missing config: Jwt:SigningKey");
-
-builder.Services
-    .AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-    .AddJwtBearer(options =>
-    {
-        options.TokenValidationParameters = new TokenValidationParameters
-        {
-            ValidateIssuer = true,
-            ValidIssuer = jwt.GetValue<string>("Issuer"),
-            ValidateAudience = true,
-            ValidAudience = jwt.GetValue<string>("Audience"),
-            ValidateLifetime = true,
-            ValidateIssuerSigningKey = true,
-            IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(signingKey)),
-            ClockSkew = TimeSpan.FromSeconds(30)
-        };
-    });
-
-builder.Services.AddAuthorization();
-
-builder.Services.AddCors(options =>
-{
-    options.AddDefaultPolicy(policy =>
-        policy.AllowAnyHeader().AllowAnyMethod().AllowAnyOrigin());
-});
+builder.Services.AddTaskFlowServices(builder.Configuration);
 
 var app = builder.Build();
-
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseCors();
-app.UseAuthentication();
-app.UseAuthorization();
-app.MapControllers();
+app.UseTaskFlowPipeline();
 
 app.Run();
