+++ backend/TaskFlow.Model/Repositories/TaskRepository.cs
@@ -15,6 +15,7 @@ public class TaskRepository : BaseRepository
     {
         var list = new List<TaskItem>();
         using var conn = new NpgsqlConnection(ConnectionString);
+
         using var cmd = new NpgsqlCommand(
             "SELECT id, title, created_at, due_at FROM tasks WHERE user_id = @user_id ORDER BY id",
             conn);
@@ -38,10 +39,11 @@ public class TaskRepository : BaseRepository
     public TaskItem Create(int userId, string title, DateTimeOffset? dueAt)
     {
         using var conn = new NpgsqlConnection(ConnectionString);
+
         using var cmd = new NpgsqlCommand(
-              @"INSERT INTO tasks (user_id, title, due_at)
-                VALUES (@user_id, @title, @due_at)
-                RETURNING id, title, created_at, due_at",
+            @"INSERT INTO tasks (user_id, title, due_at)
+              VALUES (@user_id, @title, @due_at)
+              RETURNING id, title, created_at, due_at",
             conn);
         cmd.Parameters.AddWithValue("user_id", userId);
         cmd.Parameters.AddWithValue("title", title);
@@ -60,47 +62,47 @@ public class TaskRepository : BaseRepository
             Title = reader.GetString(1),
             CreatedAt = reader.GetFieldValue<DateTimeOffset>(2),
             DueAt = reader.IsDBNull(3) ? null : reader.GetFieldValue<DateTimeOffset>(3)
-
         };
     }
 
     public TaskItem? Update(int userId, int id, string title, DateTimeOffset? dueAt)
     {
-    using var conn = new NpgsqlConnection(ConnectionString);
-    using var cmd = new NpgsqlCommand(
-        @"UPDATE tasks
-          SET title = @title,
-            due_at = @due_at
-          WHERE id = @id AND user_id = @user_id
-          RETURNING id, title, created_at, due_at",
-        conn);
+        using var conn = new NpgsqlConnection(ConnectionString);
 
-    cmd.Parameters.AddWithValue("user_id", userId);
-    cmd.Parameters.AddWithValue("id", id);
-    cmd.Parameters.AddWithValue("title", title);
+        using var cmd = new NpgsqlCommand(
+            @"UPDATE tasks
+              SET title = @title,
+                due_at = @due_at
+              WHERE id = @id AND user_id = @user_id
+              RETURNING id, title, created_at, due_at",
+            conn);
 
-    var dueAtParam = cmd.Parameters.Add("due_at", NpgsqlDbType.TimestampTz);
-    dueAtParam.Value = dueAt.HasValue ? dueAt.Value : DBNull.Value;
+        cmd.Parameters.AddWithValue("user_id", userId);
+        cmd.Parameters.AddWithValue("id", id);
+        cmd.Parameters.AddWithValue("title", title);
 
-    conn.Open();
-    using var reader = cmd.ExecuteReader();
+        var dueAtParam = cmd.Parameters.Add("due_at", NpgsqlDbType.TimestampTz);
+        dueAtParam.Value = dueAt.HasValue ? dueAt.Value : DBNull.Value;
 
-    if (!reader.Read())
-        return null;
+        conn.Open();
+        using var reader = cmd.ExecuteReader();
 
-    return new TaskItem
+        if (!reader.Read())
+            return null;
+
+        return new TaskItem
         {
-        Id = reader.GetInt32(0),
-        Title = reader.GetString(1),
-        CreatedAt = reader.GetFieldValue<DateTimeOffset>(2),
-        DueAt = reader.IsDBNull(3) ? null : reader.GetFieldValue<DateTimeOffset>(3)
-
+            Id = reader.GetInt32(0),
+            Title = reader.GetString(1),
+            CreatedAt = reader.GetFieldValue<DateTimeOffset>(2),
+            DueAt = reader.IsDBNull(3) ? null : reader.GetFieldValue<DateTimeOffset>(3)
         };
     }
 
     public bool Delete(int userId, int id)
     {
         using var conn = new NpgsqlConnection(ConnectionString);
+
         using var cmd = new NpgsqlCommand(
             "DELETE FROM tasks WHERE id = @id AND user_id = @user_id",
             conn);
